Product Admin - Securing the views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -80,10 +81,16 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """
     Add product to the store
     """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do that.')
+        return redirect(reverse("home"))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -108,10 +115,15 @@ def add_product(request):
     )
 
 
+@login_required
 def edit_product(request, product_id):
     """
     Edit a product in the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do that.')
+        return redirect(reverse("home"))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -142,10 +154,15 @@ def edit_product(request, product_id):
     )
 
 
+@login_required
 def delete_product(request, product_id):
     """
     Delete the product from the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do that.')
+        return redirect(reverse("home"))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'Product deleted!!')
